Changed webpack config for better splitting and smaller bundles
Basically two changes:

- Now all the JS for the public pages is bundled together into one entry
  point.
- All modules which are common to the public and dashboard entry points
  are pulled out into a separate 'common' chunk, which we can source of
  both pages (just increases cache-ability a little bit).

closes #1500
