+++ cms/templates/cms/home_page.html
@@ -297,7 +297,6 @@
   </div>
 </div>
 <script type="text/javascript" src="{{ public_src }}"></script>
-<script type="text/javascript" src="{{ signup_dialog_src }}"></script>
 {% endblock %}
 
 </a>div>
