+++ cms/templates/cms/program_page.html
@@ -146,9 +146,6 @@
 <div id="signup-dialog">
   <!-- signup dialog React component injected here -->
 </div>
-<script type="text/javascript" src="{{ signup_dialog_src }}"></script>
-<script type="text/javascript" src="{{ faculty_carousel_src }}"></script>
-<script type="text/javascript" src="{{ course_list_src }}"></script>
 <script type="text/javascript" src="{{ public_src }}"></script>
 <script defer src="https://code.getmdl.io/1.2.1/material.min.js"></script>
 {% endblock %}
