+++ static/js/entry/public.js
@@ -1,13 +1,67 @@
 /* eslint-disable no-unused-vars, no-undef, no-var */
 __webpack_public_path__ = `http://${SETTINGS.host}:8078/`;  // eslint-disable-line no-undef, camelcase
-
-// responsive sharing buttons
 import "rrssb/js/rrssb.js";
-
 import "bootstrap";
-
-// mailchimp requirements
 import "ajaxchimp";
+import React from 'react';
+import ReactDOM from 'react-dom';
+import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
+import getMuiTheme from 'material-ui/styles/getMuiTheme';
+import { Provider } from 'react-redux';
+import injectTapEventPlugin from 'react-tap-event-plugin';
+injectTapEventPlugin();
+
+import CourseListWithPopover from '../components/CourseListWithPopover';
+import FacultyCarousel from '../components/FacultyCarousel';
+import {
+  setDialogVisibility,
+} from '../actions/signup_dialog';
+import { signupDialogStore } from '../store/configureStore';
+import SignupDialog from '../containers/SignupDialog';
+
+// Program Page course list
+const courseListEl = document.querySelector('#course-list');
+
+if ( courseListEl ) {
+  ReactDOM.render(
+    <MuiThemeProvider muiTheme={getMuiTheme()}>
+      <CourseListWithPopover courses={SETTINGS.courses} />
+    </MuiThemeProvider>,
+    courseListEl
+  );
+}
+
+// Program Page carousel div
+const carouselDiv = document.querySelector('#faculty-carousel');
+
+if ( carouselDiv ) {
+  ReactDOM.render(
+    <FacultyCarousel faculty={SETTINGS.faculty}/>,
+    carouselDiv
+  );
+}
+
+// Signup dialog
+const store = signupDialogStore();
+
+const dialogDiv = document.querySelector('#signup-dialog');
+
+const openDialog = () => store.dispatch(setDialogVisibility(true));
+
+let nodes = [...document.querySelectorAll('.open-signup-dialog')];
+
+nodes.forEach(signUpButton => {
+  signUpButton.onclick = openDialog;
+});
+
+ReactDOM.render(
+  <MuiThemeProvider muiTheme={getMuiTheme()}>
+    <Provider store={store}>
+      <SignupDialog />
+    </Provider>
+    </MuiThemeProvider>,
+  dialogDiv
+);
 
 /// MAILCHIMP BELOW
