+++ ui/templates/base_error.html
@@ -40,5 +40,4 @@
   <!-- signup dialog React component injected here -->
   </div>
   <script type="text/javascript" src="{{ public_src }}"></script>
-  <script type="text/javascript" src="{{ signup_dialog_src }}"></script>
 {% endblock %}
