+++ ui/templates/terms_of_service.html
@@ -126,6 +126,6 @@
   <div id="signup-dialog">
     <!-- signup dialog React component injected here -->
   </div>
-  <script type="text/javascript" src="{{ signup_dialog_src }}"></script>
+  <script type="text/javascript" src="{{ public_src }}"></script>
   {% include "footer.html" %}
 {% endblock %}
